[Loom] Make source lowering plans explicit
Move source operation selection and backward materialization demand into
a source-plan component with one documented retained representation.
Rule, descriptor-matrix, and target callback decisions now complete
their planning lifecycle together before Low IR emission consumes them
in source order.

Add focused coverage through the real lowering callback surface to prove
dead-plan elision and dependency retention without encoding compiler IR
as test strings.
